Fixed ALFmxInertialMovement in iOS
i'm not sure why it occurs but on ios the DisplayLink break when scrolling (mouse down). This raise a strange silent exception that can only be viewed in xcode console log (can't handle it in app), and than the scren flicker and draw incorrect. I found some complaints about this, such as https://stackoverflow.com/questions/5944050/cadisplaylink-opengl-rendering-breaks-uiscrollview-behaviour

But this can be avoided with this simple solution that works 100% since during the touch on the screen there is no animation, the animation is only done when the touch on the screen ends.